Party.js: remove unnecessary code
the `otherParty` code made no sense. it was altogether unnecessary because anything it was doing was already implied by the `player.partyflag === 2` on the new line 138 or made no difference at all (such as the scenario where you and player were in the same party). there was also a bug in it that prevented partying in a certain situation where there are 2 different parties.