make Config.PublicMode = 1 party only with leader (#163)
* make Config.PublicMode = 1 party only with leader

the previous method had the current player accepting invites from anyone. now on `Config.PublicMode = 1` the player only accepts invites from the leader if there is one, or if there isn't one then party with anyone (same as before).

* Party.js: remove unnecessary code

the `otherParty` code made no sense. it was altogether unnecessary because anything it was doing was already implied by the `player.partyflag === 2` on the new line 138 or made no difference at all (such as the scenario where you and player were in the same party). there was also a bug in it that prevented partying in a certain situation where there are 2 different parties.