Update Necromancer.js to use CE more aggressively
This commit causes a necro to more aggressively use CE on corpses by
not trying to keep spare corpses around when the summon count is full.
Doing so vastly speeds up the clear speed on areas that are dense with
monsters and the necro already has enough summons, e.g. Cows.
For non-summoner necros the clear speed is unchanged.

Move the summon count outside of raiseArmy() into a new function
called setArmySize() that does the same thing.

Add a new function called isArmyFull() to determine if the necro has
the full amount of summons possible.
This is done in case of things like Battle Command and Skill Shrines
which would change the number needed.

If the necro has all the summons he needs he'll spam CE as if he were
a non-summoner necro, but if he is missing any then he acts the same
as before: only exploding 2 corpses at a time.