Increasing General Performance of MFHelper and al (#232)
* Optimizing the annoying delays with MFHelpers taking their time to simply change towns

1. Move the player act variable detector out of the while loop because we don't need it anymore in that main loop. We just need to query the act once to sync with the leader's starting town.

2. add a new goto command that will listen in a similar manner to the follower script.

* Introduction of intents against actuality in act changing

Just like in Mephisto.js and Andariel.js, a function is used to broadcast the leader's intent on changing acts prior to doing so. This allows the MFHelper to rapidly react to the leader's intention to change acts rather than previous implementations which relied on the leader's current act.

* Replacing skill cast with packet cast

makes sense, no?

* Packet casting TK on items

minor change, unicast is safer.

* Conditional broadcast

Broadcast only if MFLeader is true (reduces workload).

* Split replaced by substring

* Simplified getAct function

removed bin search in favor of just pop

* packet casting TK vs skill casting

TK packet casting is active if packetcasting is 2. Else, skill cast it.

* swapped while loop for polling